third_party/blink/web_tests/fast/loader/fragment-anchor-cleared-after-load-when-hidden-expected.txt

Tests that the scroll position does not stick to the anchor if it becomes visible after the page has finished loading.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS scrollY is 0
PASS successfullyParsed is true

TEST COMPLETE
